+++ src/ff7/misc.cpp
@@ -86,41 +86,6 @@ uint32_t get_equipment_stats(uint32_t party_index, uint32_t type)
 	}
 }
 
-// KERNEL2
-char *kernel2_sections[20];
-uint32_t kernel2_section_counter;
-
-void kernel2_reset_counters()
-{
-	uint32_t i;
-
-	if(trace_all) ffnx_trace("kernel2 reset\n");
-
-	for(i = 0; i < kernel2_section_counter; i++) external_free(kernel2_sections[i]);
-
-	kernel2_section_counter = 0;
-}
-
-char *kernel2_add_section(uint32_t size)
-{
-	char *ret = (char*)external_malloc(size);
-
-	if(trace_all) ffnx_trace("kernel2 add section %i (%i)\n", kernel2_section_counter, size);
-
-	kernel2_sections[kernel2_section_counter++] = ret;
-
-	return ret;
-}
-
-char *kernel2_get_text(uint32_t section_base, uint32_t string_id, uint32_t section_offset)
-{
-	char *section = kernel2_sections[section_base + section_offset];
-
-	if(trace_all) ffnx_trace("kernel2 get text (%i+%i:%i)\n", section_base, section_offset, string_id);
-
-	return &section[((WORD *)section)[string_id]];
-}
-
 // WM_ACTIVATEAPP
 void ff7_wm_activateapp(bool hasFocus)
 {